shift around checks with light
